Fixed the forms to be in accordance with these two non functional requirements:
When interacting with a text field on a form, pressing the enter key must not submit the form,
unless (potentially) when reaching the end of the form.
• When interacting with any highlightable element on the page (e.g., text fields, button),
pressing tab must move the user to the next element in an ordered manner. For example,
pressing tab move down fields on a form, but does not move the cursor randomly between
the different inputs.