FEAT: Worked on the UI for the delete button and added ability to call delete api method.
Used fetch to call an API method - this took  a while to work out because I didnt understand that the method called in fetch has to be in a RestController, which it wasn't. This lead to a lot of debugging but eventually was able to fix it. Backend (actually deleting the plant from favourites) has not been implemented - this will be implemented when the rest of this story (or atleast adding plants) has been finished as currently there aren't any repositories or services I can add to